requirements.txt
# Core Dependencies
llama-index>=0.10.0
groq>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.3.1
transformers>=4.38.0
faker>=20.0.0  # For synthetic data generation
pystemmer>=2.2.0  # For BM25 retriever
chromadb>=0.4.0  # For vector storage
streamlit>=1.30.0  # For UI
threading>=0.0.0  # For IoT simulation

# PDF Processing
unstructured>=0.12.0
pypdf>=4.2.0
pdfminer.six>=20221105
pdfplumber>=0.10.0

# IoT Integration
thingspeak>=1.0.0  # For IoT data integration
paho-mqtt>=2.0.0  # Optional for MQTT support

# Additional Utilities
pydantic>=2.5.0  # For data validation
loguru>=0.7.0  # Enhanced logging
tqdm>=4.66.0  # Progress bars
python-dateutil>=2.8.2  # Date handling
scipy>=1.11.0  # For advanced data analysis
scikit-learn>=1.3.0  # For BM25 retriever